Membangun_model/requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
mlflow==2.8.1
numpy==1.24.3
dagshub==0.3.12
python-dotenv==1.0.1